[lit] Split benchmark.py into a package
The single-file benchmark script had bloated ~1250 lines, tangling up utilities, config, platform strategies, runner logic, and CLI parsing. Breaking it down into dedicated modules (util, config, runner, cli, platforms/) improves readability and makes each piece easier to test, without altering any underlying behavior.
