+++ llvm/utils/lit/utils/benchmark/__main__.py
@@ -1,0 +1,8 @@
+"""Enable `python -m benchmark`."""
+
+from __future__ import annotations
+
+from benchmark.cli import main
+
+if __name__ == "__main__":
+    main()
